requirements.txt
# 백엔드 동작에 필요한 라이브러리

django==4.1.5
djangorestframework==3.14.0
drf-spectacular==0.25.1
pygments==2.14.0
gunicorn==20.1.0

# YOLOv5 동작 과정에서 필요한 라이브러리

matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
seaborn>=0.11.0
pandas
dill>=0.3.7
# 추가 --------------------------------------
ipython  
psutil  
thop>=0.1.1  
